server/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.31
uvicorn==0.30.1
pydantic-settings
alembic
asyncpg
psycopg2